Create new wrapper component, in prep for merging
