+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/insert-property-row.tsx
@@ -5,18 +5,16 @@ import {
   TableRow,
   useForkRef,
 } from "@mui/material";
-import {
-  bindToggle,
-  bindTrigger,
-  usePopupState,
-} from "material-ui-popup-state/hooks";
+import { bindTrigger, usePopupState } from "material-ui-popup-state/hooks";
 import { Ref, useId, useRef, useState } from "react";
 import { useFormContext, useWatch } from "react-hook-form";
 import { useBlockProtocolCreatePropertyType } from "../../../../components/hooks/blockProtocolFunctions/ontology/useBlockProtocolCreatePropertyType";
 import { EntityTypeEditorForm } from "./form-types";
 import { PropertyTypeSelector } from "./property-type-selector";
-import { PropertyTypeForm, PropertyTypeFormValues } from "./property-type-form";
-import { PropertyTypeModal } from "./property-type-modal";
+import {
+  PropertyTypeFormValues,
+  PropertyTypeModalForm,
+} from "./property-type-modal-form";
 import { QuestionIcon } from "./question-icon";
 import { useRefetchPropertyTypes } from "./use-property-types";
 import { useRouteNamespace } from "./use-route-namespace";
@@ -112,9 +110,8 @@ export const InsertPropertyRow = ({
             variant: "propertyType",
           }}
         />
-        <PropertyTypeModal
-          popupState={modalPopupState}
-          title={
+        <PropertyTypeModalForm
+          modalTitle={
             <>
               Create new property type
               <QuestionIcon
@@ -124,17 +121,11 @@ export const InsertPropertyRow = ({
               />
             </>
           }
-        >
-          <PropertyTypeForm
-            discardButtonProps={withHandler(
-              bindToggle(modalPopupState),
-              () => {},
-            )}
-            initialTitle={searchText}
-            onSubmit={handleSubmit}
-            submitButtonProps={{ children: <>Create new property type</> }}
-          />
-        </PropertyTypeModal>
+          popupState={modalPopupState}
+          onSubmit={handleSubmit}
+          submitButtonProps={{ children: <>Create new property type</> }}
+          initialTitle={searchText}
+        />
       </TableCell>
     </TableRow>
   );
